# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PalPanel API
  version: 1.0.0
  description: Stable HTTP contract for the PalPanel 1.x management API.
servers:
  - url: /api
security:
  - sessionCookie: []
  - developmentKey: []
...
paths:
...
  /bases/{id}/name:
    parameters: [*idParameter]
    put: {operationId: updateBaseCustomName, tags: [world], x-palpanel-permission: 'server:control', requestBody: *jsonRequest, responses: *writeResponses}
    delete: {operationId: deleteBaseCustomName, tags: [world], x-palpanel-permission: 'server:control', responses: *writeResponses}
...
  /inventory:
    get:
      operationId: listGlobalInventory
      tags: [world]
      x-palpanel-permission: read
      parameters:
        - {name: q, in: query, schema: {type: string}}
        - {name: category, in: query, schema: {type: string}}
        - {name: owner_type, in: query, schema: {type: string, enum: [all, base, player, unknown], default: all}}
        - {name: sort, in: query, schema: {type: string, enum: [count_desc, name_asc, name_desc], default: count_desc}}
        - {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 500, default: 200}}
        - {name: offset, in: query, schema: {type: integer, minimum: 0, default: 0}}
      responses: *readResponses
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: palpanel_session
      description: HttpOnly SameSite=Lax session cookie created by registration or login.
    developmentKey:
      type: http
      scheme: bearer
      bearerFormat: ppk_...
      description: 'Revocable development key supplied as Authorization: Bearer ppk_...'
...